requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.31.6
MarkupSafe==2.1.5
msgpack==1.1.0
packaging==24.1
playwright==1.45.1
pluggy==1.5.0
psutil==6.0.0
pycparser==2.22
pyee==11.1.0
pytest==8.3.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.5.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pywin32==306
pyzmq==26.2.0
requests==2.32.3
setuptools==75.1.0
six==1.16.0
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
